requirements.txt
cryptography==3.3.1
pycrypto==2.6.1
pycryptodome==3.9.9
pycryptodomex==3.9.9
streamlit==0.74.1
dependency-injector==4.10.0